+++ lib/admob_easy.dart
@@ -3,4 +3,5 @@ library admob_easy;
 export 'package:admob_easy/ads/admob_easy.dart';
 export 'ads/services/banner_ad.dart';
 export 'ads/services/open_app_ad.dart';
+export 'ads/services/admob_easy_native.dart';
 export 'package:google_mobile_ads/google_mobile_ads.dart';